Creating basic DAO for pets

+++ pet_project_jar/src/main/java/org/home/dao/UserDAO.java
@@ -2,6 +2,6 @@ package org.home.dao;
 
 import org.home.entities.User;
 
-public class UserDAO extends GenericDAO<User, Integer> {
+public class UserDAO extends GenericDAO<User, Long> {
 
 }
